Pipeline/requirements.txt
pandas
pyyaml
scikit-learn
lightgbm
fastapi
uvicorn
